backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Configuration management
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# HTTP client for web scraping
httpx==0.25.2
aiohttp==3.9.1

# Data processing
pandas==2.1.4
openpyxl==3.1.2

# Image processing
Pillow==10.1.0

# Supabase
supabase==2.3.4

# Translation (for future tasks)
openai==1.3.7

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0

# HTTP requests
requests==2.31.0

# Type stubs
types-redis==4.6.0.11
types-requests==2.31.0.10

# CORS middleware
python-multipart==0.0.6
